Accept Hash for group children
A valid Ansible inventory would contain a Hash in a groups children
value. So far we only used lists there and converted it to a hash. Now
we intend to do group nesting and the code faile due to product not
being a valid method on a hash. As there is nothing to do in this case,
one more test for type Array prevents the code from failing.